Reorganized tests around scheme/format concept (#61)
