+++ tests/formats/test_ods.py
@@ -5,14 +5,33 @@ from __future__ import absolute_import
 from __future__ import unicode_literals
 
 import io
+import pytest
 from mock import Mock
-from tabulator import Stream
+from tabulator import Stream, exceptions
 from tabulator.parsers.ods import ODSParser
+BASE_URL = 'https://raw.githubusercontent.com/okfn/tabulator-py/master/%s'
 
 
-# Tests
+# Stream
 
-def test_ods_parser():
+def test_stream_ods():
+    with Stream('data/table.ods', headers=1) as stream:
+        assert stream.headers == ['id', 'name']
+        assert stream.read(keyed=True) == [
+            {'id': 1.0, 'name': 'english'},
+            {'id': 2.0, 'name': '中国人'},
+        ]
+
+
+def test_stream_ods_remote():
+    source = BASE_URL % 'data/table.ods'
+    with Stream(source) as stream:
+        assert stream.read() == [['id', 'name'], [1.0, 'english'], [2.0, '中国人']]
+
+
+# Parser
+
+def test_parser_ods():
 
     source = 'data/table.ods'
     encoding = None
