+++ tests/formats/test_xls.py
@@ -7,12 +7,34 @@ from __future__ import unicode_literals
 import io
 from mock import Mock
 from tabulator import parsers
+from tabulator import Stream, exceptions
 from tabulator.parsers.xls import XLSParser
+BASE_URL = 'https://raw.githubusercontent.com/okfn/tabulator-py/master/%s'
 
 
-# Tests
+# Stream
 
-def test_xls_parser():
+def test_stream_local_xls():
+    with Stream('data/table.xls') as stream:
+        assert stream.headers is None
+        assert stream.read() == [['id', 'name'], [1.0, 'english'], [2.0, '中国人']]
+
+
+def test_stream_remote_xls():
+    with Stream(BASE_URL % 'data/table.xls') as stream:
+        assert stream.headers is None
+        assert stream.read() == [['id', 'name'], [1.0, 'english'], [2.0, '中国人']]
+
+
+def test_stream_xls_sheet():
+    source = 'data/special/sheet2.xls'
+    with Stream(source, sheet=2) as stream:
+        assert stream.read() == [['id', 'name'], [1.0, 'english'], [2.0, '中国人']]
+
+
+# Parser
+
+def test_parser_xls():
 
     source = 'data/table.xls'
     encoding = None
