--- tests/parsers/test_csv.py
@@ -1,36 +1,0 @@
-# -*- coding: utf-8 -*-
-from __future__ import division
-from __future__ import print_function
-from __future__ import absolute_import
-from __future__ import unicode_literals
-
-import io
-from mock import Mock
-from tabulator.parsers.csv import CSVParser
-
-
-# Tests
-
-def test_csv_parser():
-
-    source = 'data/table.csv'
-    encoding = None
-    loader = Mock()
-    loader.load = Mock(return_value=io.open(source, encoding='utf-8'))
-    parser = CSVParser(loader)
-
-    assert parser.closed
-    parser.open(source, encoding=encoding)
-    assert not parser.closed
-
-    assert list(parser.extended_rows) == [
-        (1, None, ['id', 'name']),
-        (2, None, ['1', 'english']),
-        (3, None, ['2', '中国人'])]
-
-    assert len(list(parser.extended_rows)) == 0
-    parser.reset()
-    assert len(list(parser.extended_rows)) == 3
-
-    parser.close()
-    assert parser.closed
