+++ tests/test_stream.py
@@ -14,7 +14,6 @@ from tabulator import Stream, exceptions
 from tabulator.loaders.local import LocalLoader
 from tabulator.parsers.csv import CSVParser
 from tabulator.writers.csv import CSVWriter
-BASE_URL = 'https://raw.githubusercontent.com/okfn/tabulator-py/master/%s'
 
 
 # Headers
@@ -82,335 +81,6 @@ def test_stream_headers_inline_keyed_headers_is_none():
         assert list(stream.iter(extended=True)) == [
             (1, None, ['1', 'english']),
             (2, None, ['2', '中国人'])]
-
-
-# Scheme: local
-
-def test_stream_local():
-    with Stream('data/table.csv') as stream:
-        assert stream.read() == [['id', 'name'], ['1', 'english'], ['2', '中国人']]
-
-
-# Scheme: remote
-
-def test_stream_remote():
-    with Stream(BASE_URL % 'data/table.csv') as stream:
-        assert stream.read() == [['id', 'name'], ['1', 'english'], ['2', '中国人']]
-
-
-# Scheme: stream
-
-def test_stream_stream():
-    source = io.open('data/table.csv', mode='rb')
-    with Stream(source, format='csv') as stream:
-        assert stream.read() == [['id', 'name'], ['1', 'english'], ['2', '中国人']]
-
-
-# Scheme: text
-
-def test_stream_text():
-    source = 'text://value1,value2\nvalue3,value4'
-    with Stream(source, format='csv') as stream:
-        assert stream.read() == [['value1', 'value2'], ['value3', 'value4']]
-
-
-# Format: csv
-
-def test_stream_local_csv():
-    with Stream('data/table.csv') as stream:
-        assert stream.headers is None
-        assert stream.read() == [['id', 'name'], ['1', 'english'], ['2', '中国人']]
-
-
-def test_stream_local_csv_with_bom():
-    with Stream('data/special/bom.csv') as stream:
-        assert stream.headers is None
-        assert stream.read() == [['id', 'name'], ['1', 'english'], ['2', '中国人']]
-
-
-def test_stream_local_csv_with_bom_with_encoding():
-    with Stream('data/special/bom.csv', encoding='utf-8') as stream:
-        assert stream.headers is None
-        assert stream.read() == [['id', 'name'], ['1', 'english'], ['2', '中国人']]
-
-
-def test_stream_csv_excel():
-    source = 'value1,value2\nvalue3,value4'
-    with Stream(source, scheme='text', format='csv') as stream:
-        assert stream.read() == [['value1', 'value2'], ['value3', 'value4']]
-
-
-def test_stream_csv_excel_tab():
-    source = 'value1\tvalue2\nvalue3\tvalue4'
-    with Stream(source, scheme='text', format='csv', delimiter='\t') as stream:
-        assert stream.read() == [['value1', 'value2'], ['value3', 'value4']]
-
-
-def test_stream_csv_unix():
-    source = '"value1","value2"\n"value3","value4"'
-    with Stream(source, scheme='text', format='csv') as stream:
-        assert stream.read() == [['value1', 'value2'], ['value3', 'value4']]
-
-
-def test_stream_csv_escaping():
-    with Stream('data/special/escaping.csv', escapechar='\\') as stream:
-        assert stream.read() == [
-            ['ID', 'Test'],
-            ['1', 'Test line 1'],
-            ['2', 'Test " line 2'],
-            ['3', 'Test " line 3'],
-        ]
-
-
-def test_stream_csv_doublequote():
-    with Stream('data/special/doublequote.csv') as stream:
-        for row in  stream:
-            assert len(row) == 17
-
-
-def test_stream_stream_csv():
-    source = io.open('data/table.csv', mode='rb')
-    with Stream(source, format='csv') as stream:
-        assert stream.headers is None
-        assert stream.read() == [['id', 'name'], ['1', 'english'], ['2', '中国人']]
-
-
-def test_stream_text_csv():
-    source = 'text://id,name\n1,english\n2,中国人\n'
-    with Stream(source, format='csv') as stream:
-        assert stream.headers is None
-        assert stream.read() == [['id', 'name'], ['1', 'english'], ['2', '中国人']]
-
-
-def test_stream_remote_csv():
-    with Stream(BASE_URL % 'data/table.csv') as stream:
-        assert stream.headers is None
-        assert stream.read() == [['id', 'name'], ['1', 'english'], ['2', '中国人']]
-
-
-def test_stream_remote_csv_non_ascii_url():
-    with Stream('http://data.defra.gov.uk/ops/government_procurement_card/over_£500_GPC_apr_2013.csv') as stream:
-        assert stream.sample[0] == [
-            'Entity',
-            'Transaction Posting Date',
-            'Merchant Name',
-            'Amount',
-            'Description']
-
-
-def test_stream_csv_delimiter():
-    source = '"value1";"value2"\n"value3";"value4"'
-    with Stream(source, scheme='text', format='csv', delimiter=';') as stream:
-        assert stream.read() == [['value1', 'value2'], ['value3', 'value4']]
-
-
-def test_stream_csv_escapechar():
-    source = 'value1%,value2\nvalue3%,value4'
-    with Stream(source, scheme='text', format='csv', escapechar='%') as stream:
-        assert stream.read() == [['value1,value2'], ['value3,value4']]
-
-
-def test_stream_csv_quotechar():
-    source = '%value1,value2%\n%value3,value4%'
-    with Stream(source, scheme='text', format='csv', quotechar='%') as stream:
-        assert stream.read() == [['value1,value2'], ['value3,value4']]
-
-
-def test_stream_csv_quotechar():
-    source = 'value1, value2\nvalue3, value4'
-    with Stream(source, scheme='text', format='csv', skipinitialspace=True) as stream:
-        assert stream.read() == [['value1', 'value2'], ['value3', 'value4']]
-
-
-# Format: gsheet
-
-def test_stream_gsheet():
-    source = 'https://docs.google.com/spreadsheets/d/1mHIWnDvW9cALRMq9OdNfRwjAthCUFUOACPp0Lkyl7b4/edit?usp=sharing'
-    with Stream(source) as stream:
-        assert stream.read() == [['id', 'name'], ['1', 'english'], ['2', '中国人']]
-
-
-def test_stream_gsheet_with_gid():
-    source = 'https://docs.google.com/spreadsheets/d/1mHIWnDvW9cALRMq9OdNfRwjAthCUFUOACPp0Lkyl7b4/edit#gid=960698813'
-    with Stream(source) as stream:
-        assert stream.read() == [['id', 'name'], ['2', '中国人'], ['3', 'german']]
-
-
-def test_stream_gsheet_bad_url():
-    stream = Stream('https://docs.google.com/spreadsheets/d/bad')
-    with pytest.raises(exceptions.HTTPError) as excinfo:
-        stream.open()
-
-
-# Format: inline
-
-def test_stream_inline():
-    source = [['id', 'name'], ['1', 'english'], ['2', '中国人']]
-    with Stream(source) as stream:
-        assert stream.headers is None
-        assert stream.read() == [['id', 'name'], ['1', 'english'], ['2', '中国人']]
-
-
-def test_stream_inline_iterator():
-    source = iter([['id', 'name'], ['1', 'english'], ['2', '中国人']])
-    with Stream(source) as stream:
-        assert stream.headers is None
-        assert stream.read() == [['id', 'name'], ['1', 'english'], ['2', '中国人']]
-
-
-def test_stream_inline_iterator():
-    def generator():
-        yield ['id', 'name']
-        yield ['1', 'english']
-        yield ['2', '中国人']
-    with pytest.raises(exceptions.SourceError) as excinfo:
-        iterator = generator()
-        Stream(iterator).open()
-    assert 'callable' in str(excinfo.value)
-
-
-def test_stream_inline_generator():
-    def generator():
-        yield ['id', 'name']
-        yield ['1', 'english']
-        yield ['2', '中国人']
-    with Stream(generator) as stream:
-        assert stream.headers is None
-        assert stream.read() == [['id', 'name'], ['1', 'english'], ['2', '中国人']]
-
-
-def test_stream_inline_keyed():
-    source = [{'id': '1', 'name': 'english'}, {'id': '2', 'name': '中国人'}]
-    with Stream(source, format='inline') as stream:
-        assert stream.headers is None
-        assert stream.read() == [['1', 'english'], ['2', '中国人']]
-
-
-# Format: json
-
-def test_stream_local_json_dicts():
-    with Stream('data/table-dicts.json') as stream:
-        assert stream.headers is None
-        assert stream.read() == [[1, 'english'], [2, '中国人']]
-
-
-def test_stream_local_json_lists():
-    with Stream('data/table-lists.json') as stream:
-        assert stream.headers is None
-        assert stream.read() == [['id', 'name'], [1, 'english'], [2, '中国人']]
-
-
-def test_stream_text_json_dicts():
-    source = '[{"id": 1, "name": "english" }, {"id": 2, "name": "中国人" }]'
-    with Stream(source, scheme='text', format='json') as stream:
-        assert stream.headers is None
-        assert stream.read() == [[1, 'english'], [2, '中国人']]
-
-
-def test_stream_text_json_lists():
-    source = '[["id", "name"], [1, "english"], [2, "中国人"]]'
-    with Stream(source, scheme='text', format='json') as stream:
-        assert stream.headers is None
-        assert stream.read() == [['id', 'name'], [1, 'english'], [2, '中国人']]
-
-
-def test_stream_remote_json_dicts():
-    with Stream(BASE_URL % 'data/table-dicts.json') as stream:
-        assert stream.headers is None
-        assert stream.read() == [[1, 'english'], [2, '中国人']]
-
-
-def test_stream_remote_json_lists():
-    with Stream(BASE_URL % 'data/table-lists.json') as stream:
-        assert stream.headers is None
-        assert stream.read() == [['id', 'name'], [1, 'english'], [2, '中国人']]
-
-
-# Format: ods
-
-def test_stream_ods():
-    with Stream('data/table.ods', headers=1) as stream:
-        assert stream.headers == ['id', 'name']
-        assert stream.read(keyed=True) == [
-            {'id': 1.0, 'name': 'english'},
-            {'id': 2.0, 'name': '中国人'},
-        ]
-
-
-def test_stream_ods_remote():
-    source = BASE_URL % 'data/table.ods'
-    with Stream(source) as stream:
-        assert stream.read() == [['id', 'name'], [1.0, 'english'], [2.0, '中国人']]
-
-
-# Format: sql
-
-def test_stream_format_sql(database_url):
-    with Stream(database_url, table='data') as stream:
-        assert stream.read() == [[1, 'english'], [2, '中国人']]
-
-
-def test_stream_format_sql_order_by(database_url):
-    with Stream(database_url, table='data', order_by='id') as stream:
-        assert stream.read() == [[1, 'english'], [2, '中国人']]
-
-
-def test_stream_format_sql_order_by_desc(database_url):
-    with Stream(database_url, table='data', order_by='id desc') as stream:
-        assert stream.read() == [[2, '中国人'], [1, 'english']]
-
-
-def test_stream_format_sql_table_is_required_error(database_url):
-    with pytest.raises(exceptions.OptionsError) as excinfo:
-        Stream(database_url).open()
-    assert 'table' in str(excinfo.value)
-
-
-def test_stream_format_sql_headers(database_url):
-    with Stream(database_url, table='data', headers=1) as stream:
-        assert stream.headers == ['id', 'name']
-        assert stream.read() == [[1, 'english'], [2, '中国人']]
-
-
-# Format: xls
-
-def test_stream_local_xls():
-    with Stream('data/table.xls') as stream:
-        assert stream.headers is None
-        assert stream.read() == [['id', 'name'], [1.0, 'english'], [2.0, '中国人']]
-
-
-def test_stream_remote_xls():
-    with Stream(BASE_URL % 'data/table.xls') as stream:
-        assert stream.headers is None
-        assert stream.read() == [['id', 'name'], [1.0, 'english'], [2.0, '中国人']]
-
-
-def test_stream_xls_sheet():
-    source = 'data/special/sheet2.xls'
-    with Stream(source, sheet=2) as stream:
-        assert stream.read() == [['id', 'name'], [1.0, 'english'], [2.0, '中国人']]
-
-
-# Format: xlsx
-
-def test_stream_xlsx_remote():
-    source = BASE_URL % 'data/table.xlsx'
-    with Stream(source) as stream:
-        assert stream.read() == [['id', 'name'], [1.0, 'english'], [2.0, '中国人']]
-
-
-def test_stream_stream_xlsx():
-    source = io.open('data/table.xlsx', mode='rb')
-    with Stream(source, format='xlsx') as stream:
-        assert stream.headers is None
-        assert stream.read() == [['id', 'name'], [1.0, 'english'], [2.0, '中国人']]
-
-
-def test_stream_xlsx_sheet():
-    source = 'data/special/sheet2.xlsx'
-    with Stream(source, sheet=2) as stream:
-        assert stream.read() == [['id', 'name'], [1, 'english'], [2, '中国人']]
 
 
 # Encoding
